requirements.txt
flask
opencv-python
tensorflow
tensorflow-keras
kears-tf 
pillow
numpy
kagglehub
matplotlib
